[폴더위계정리] refactor: ModelExt.kt dead code 제거 및 역할별 분해
DTO 의존 함수(LectureDto.contains, LectureDto.getFittingTrimParam)와
미사용 함수(getReviewUrl) 삭제 후 역할별 파일로 분리:
- CourseBookExt.kt: CourseBook 표시 문자열
- TagTypeExt.kt: TagType → 색상 매핑
- TimetableExt.kt: 시간표 trim/fit/contains
- concatenate → List + 연산자로 인라인
- toCreditNumber → LectureSearchPagingSource private 함수로 이동
